tesseract-ocr/pyproject.toml
[project]
name = "tesseract-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.3",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
]
